Add firebase integration and exception handler (#3)
* add firebase integration, error handling

* update prod profile properties

* bad request handler

+++ src/main/java/com/postgresql/Daruma/repo/ColorRepo.java
@@ -1,9 +1,10 @@
 package com.postgresql.Daruma.repo;
 
-import com.postgresql.Daruma.model.Color;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.postgresql.Daruma.model.Color;
+
 @Repository
-public interface ColorRepo extends JpaRepository<Color, Long> {
+public interface ColorRepo extends JpaRepository<Color, Integer> {
 }
