+++ src/main/java/com/postgresql/Daruma/repo/DarumaRepo.java
@@ -1,9 +1,10 @@
 package com.postgresql.Daruma.repo;
 
-import com.postgresql.Daruma.model.Daruma;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.postgresql.Daruma.model.Daruma;
+
 @Repository
-public interface DarumaRepo extends JpaRepository<Daruma, Long> {
+public interface DarumaRepo extends JpaRepository<Daruma, Integer> {
 }
